requirements.txt
Flask==3.0.3
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.4.2
joblib==1.4.2
python-dotenv==1.0.1
groq==0.9.0
matplotlib==3.8.4
reportlab==4.2.2